pyproject.toml
[project]
name = "coppermind-rag"
version = "0.1.0"
description = "RAG based on the coppermind wiki"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.38.44",
    "python-dotenv>=1.1.1",
]
